Fixing the CSS and adding a logic of a separate function for fade up effect
